+++ sources/Panel/src/Hardware/HAL/HAL_FSMC.cpp
@@ -272,7 +272,7 @@ void HAL_BUS_DISPLAY::SendBuffer(uint8 *buffer, int x, int y, int width, int hei
         uint16 col3 = (uint16)color2;
         uint16 value = (uint16)(color1 << 8);
         __asm ( "nop" );
-        PORT_DATA->ODR = value | (uint16)(color1 >> 8);
+        PORT_DATA->ODR = (uint)(value | (uint16)(color1 >> 8));
         PORT_WR->BSRR = PIN_WR;
 
         value |= (uint8)(color2 >> 16);
